Fix for RootCloak causing "Unsatisfied Link Error"
When Root Cloak is active, the check for RootNative is not given access
to a certain set of DLL’s (of which I don’t know). While these
Libraries can be loaded into memory, the application will be denied
write access.

I changed the “checkForRootNative()” method to catch this error and
return true because this error seems to only occur when the device is
rooted and has RootCloak active.

Further development is nessisary to transition the check into the
“detectRootCloakingApps()” method because the method is returning a
false negative.